refactor: reorder component folder structure
