+++ frontend/app/(auth-pages)/sign-in/page.tsx
@@ -2,8 +2,8 @@
 
 import { signInAction } from "@/app/actions";
 
-import { FormMessage, Message } from "@/components/form-message";
-import { SubmitButton } from "@/components/submit-button";
+import { FormMessage, Message } from "@/components/auth/form-message";
+import { SubmitButton } from "@/components/ui/submit-button";
 import { Input } from "@/components/ui/input";
 import { Form, FormControl, FormField, FormItem, FormLabel } from "@/components/ui/form";
