+++ frontend/components/auth/header-auth.tsx
@@ -1,7 +1,7 @@
 import { hasEnvVars } from "@/utils/supabase/check-env-vars";
 import Link from "next/link";
-import { Badge } from "./ui/badge";
-import { Button } from "./ui/button";
+import { Badge } from "../ui/badge";
+import { Button } from "../ui/button";
 import { createClient } from "@/utils/supabase/server";
 import SignOutButton from "./sign-out-button";
